Improve valuation estimate, to avoid stacking up at boundaries
Original valuation algorithm chooses an estimate from a uniform
distribution over [true_value - valuation_inaccuracy, true_value + valuation_inaccuracy]
and then pins the result to fall in the range [0, 100]. This results in
a lot of estimates of 0 or 100 exactly, for high valuation_inaccuracy values.

New valuation algorithm forces the valuation distribution to be uniform
over a sub-interval of [10, 100] instead. The lower bound of 10 is chosen
because the opportunity identification dimension will zero out some
estimates already.